Spatialization!
Now mono sources will pan and attenuate by distance model. This is a lot of
math I don't understand well and spent a lot of time researching, so be
gentle!

It's not clear to me that Constant Power Panning, and my hack to split it
into quadrants, is a good idea. There are possibly better approaches,
possibly not using Constant Power Panning at all.

Obviously this is a stereo thing. Surround sound needs better hacks.  :)